Added requirejs-config.js to allow module to work without depending on other modules.
This fixes Monetra module breaking unless either Magento_Paypal or Magento_Authorizenet are enabled.  However, neither of those modules are listed as a dependency.
Without either of those modules enabled transparent.js is not available in the static files and checkout fails.

By adding this requirejs-config.js no other payment modules are required to be enabled.